+++ Date.cpp
@@ -1,7 +1,6 @@
 // For review purposes only! Does not represent the actual solution!
 
-#include "date.h"
-#include <string>
+#include "Date.h"
 
 Date::Date()
 {
