+++ Regist.cpp
@@ -1,6 +1,6 @@
 // For review purposes only! Does not represent the actual solution!
 
-#include "regist.h"
+#include "Regist.h"
 
 Registration::Registration()
 {
