+++ Unit.cpp
@@ -1,7 +1,6 @@
 // For review purposes only! Does not represent the actual solution!
 
-#include "unit.h"
-#include <string>
+#include "Unit.h"
 
 Unit::Unit()
 {
